requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.26.0
gspread>=6.0.0
google-auth>=2.29.0
requests>=2.31.0
python-dotenv>=1.0.0
